refactor: 명명 수정
