+++ server/src/main/java/com/ahmadda/application/EventUpdateListener.java
@@ -1,8 +1,7 @@
-package com.ahmadda.application.event;
+package com.ahmadda.application;
 
+import com.ahmadda.application.dto.EventUpdated;
 import com.ahmadda.application.exception.NotFoundException;
-import com.ahmadda.domain.Event;
-import com.ahmadda.domain.EventRepository;
 import com.ahmadda.domain.EventStatistic;
 import com.ahmadda.domain.EventStatisticRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,17 +14,16 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class EventCreatedListener {
+public class EventUpdateListener {
 
     private final EventStatisticRepository eventStatisticRepository;
-    private final EventRepository eventRepository;
 
     @EventListener
     @Transactional
-    public void onEventCreated(final EventCreated eventCreated) {
-        Event event = eventRepository.findById(eventCreated.eventId())
-                .orElseThrow(() -> new NotFoundException("존재하지 않는 이벤트입니다."));
+    public void onEventUpdated(final EventUpdated eventUpdated) {
+        EventStatistic eventStatistic = eventStatisticRepository.findByEventId(eventUpdated.eventId())
+                .orElseThrow(() -> new NotFoundException("해당되는 이벤트의 조회수를 가져오는데 실패하였습니다."));
 
-        eventStatisticRepository.save(EventStatistic.create(event));
+        eventStatistic.updateEventViewMatricUntilEventEnd();
     }
 }
